feat(modal): extract style and helpers into their own files
affects: @crave/farmblocks-modal
