+++ .eslintrc.js
@@ -42,4 +42,7 @@ module.exports = {
       },
     },
   ],
+  env: {
+    browser: true,
+  },
 };
