+++ packages/modal/src/Modal.js
@@ -1,39 +1,10 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect } from "react";
 import ReactDOM from "react-dom";
-/* eslint-disable react/prop-types */
-// eslint-disable-next-line no-unused-vars
 import PropTypes from "prop-types";
-import styled from "styled-components";
-import { colors } from "@crave/farmblocks-theme";
-import Card from "@crave/farmblocks-card";
 
-const Overlay = styled.div`
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background: ${colors.GREY_32};
-  padding: 40px;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-`;
+import { Overlay, ConstrainedCard, ContentWrapper } from "./Modal.styled";
 
-const ConstrainedCard = styled(Card)`
-  margin: 0 auto;
-  max-height: 100%;
-  width: auto;
-  max-width: 100%;
-`;
-
-const ContentWrapper = styled.div`
-  height: 100%;
-  overflow: auto;
-  -webkit-overflow-scrolling: touch;
-`;
-
-export const Modal = ({
+const Modal = ({
   isOpen,
   parentNode,
   shouldCloseOnOverlayClick,
@@ -78,36 +49,18 @@ export const Modal = ({
   );
 };
 Modal.defaultProps = {
+  isOpen: false,
   parentNode: document.body,
   shouldCloseOnOverlayClick: true,
   shouldCloseOnEsc: true,
 };
-
-export const useModal = () => {
-  const [isOpen, setOpen] = useState(false);
-  const actions = {
-    open: () => setOpen(true),
-    close: () => setOpen(false),
-    toggle: () => setOpen(open => !open),
-  };
-  const Comp = ({ onRequestClose, ...props }) => (
-    <Modal
-      isOpen={isOpen}
-      onRequestClose={(...args) => {
-        actions.close();
-        onRequestClose?.(...args);
-      }}
-      {...props}
-    />
-  );
-
-  return [Comp, actions];
-};
-
-export const ModalFactory = ({ children }) => {
-  const [ModalComp, actions] = useModal();
-
-  return children(ModalComp, actions);
+Modal.propTypes = {
+  isOpen: PropTypes.bool,
+  parentNode: PropTypes.instanceOf(HTMLElement),
+  shouldCloseOnOverlayClick: PropTypes.bool,
+  shouldCloseOnEsc: PropTypes.bool,
+  onRequestClose: PropTypes.func,
+  children: PropTypes.node,
 };
 
 export default Modal;
