+++ packages/modal/src/helpers.js
@@ -1,0 +1,27 @@
+import React, { useState } from "react";
+
+import Modal from "./Modal";
+
+export const useModal = () => {
+  const [isOpen, setOpen] = useState(false);
+  const actions = {
+    open: () => setOpen(true),
+    close: () => setOpen(false),
+    toggle: () => setOpen(open => !open),
+  };
+  const Comp = ({ onRequestClose, ...props }) => (
+    <Modal
+      isOpen={isOpen}
+      onRequestClose={(...args) => {
+        actions.close();
+        onRequestClose?.(...args);
+      }}
+      {...props}
+    />
+  );
+  Comp.propTypes = Modal.propTypes;
+
+  return [Comp, actions];
+};
+
+export const ModalFactory = ({ children }) => children(...useModal());
